refactor(docker): move all docker structs in the same directory
