+++ edgehog-device-runtime-docker/src/docker/image.rs
