+++ edgehog-device-runtime-docker/src/lib.rs
@@ -28,12 +28,9 @@
 pub(crate) mod client;
 pub mod docker;
 pub mod error;
-pub mod image;
-pub mod network;
 mod properties;
 pub mod request;
 pub mod service;
-pub mod volume;
 
 #[cfg(feature = "mock")]
 mod mock;
